api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pymongo==4.6.0
python-dotenv==1.0.0
mangum==0.17.0
